+++ frontend/src/plugins/impl/data-frames/forms/form.tsx
@@ -41,7 +41,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-import { DataTypeIcon } from "./DataTypeIcon";
+import { DataTypeIcon } from "./datatype-icon";
 
 interface Props<T extends FieldValues> {
   form: UseFormReturn<T>;
